requirements.txt
asn1crypto==0.22.0
attrs==15.2.0
Automat==0.0.0
cffi==1.10.0
constantly==15.1.0
cryptography==1.8.1
cssselect==1.0.1
idna==2.5
incremental==16.10.1
lxml==3.7.3
packaging==16.8
parsel==1.1.0
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycparser==2.17
PyDispatcher==2.0.5
pyOpenSSL==17.0.0
pyparsing==2.1.4
pywin32==220
queuelib==1.4.2
Scrapy==1.3.3
service-identity==16.0.0
six==1.10.0
Twisted==17.1.0
w3lib==1.17.0
zope.interface==4.4.1
selenium==3.5.0
python-dateutil==2.6.1